moved loadGithubIssues and listRepos into the GithubIntergration class
added a function to check if the render service is awake and to start the service.

added an extra check to prevent github issue to not load if the service has been spin down
